requirements.txt
PySide6>=6.6.0
mss>=9.0.1
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
scikit-image>=0.22.0
